Remote or Local resolution of Bash dependencies.
Allows the user to use locally stored versions of the dependent files, in the event that they are executing the manager locally (from within the Git/Source directory).

This will be useful in the future if this script is expanded in any ways where it may be installed locally in /usr/bin for managing the current installation.

That said, in that case, we may have to split manager and installer, but, whateves.  The research and groundwork is there anyway.