Increase Watchdog timer per documentation recomendations.
The documentation for SdNotify make it very clear that we should be notifying the watchdog at least time/2 (which is to say [WatchdogSec/2].

The primary loop that has done this was `UI.update` although I see that additional calls were added to try to resolve the random reboots.  The fact of the matter is, the timer was 10 seconds, and update was only called every 10 seconds...  So...  A delay of even a second, anywhere in the app, and the watchdog would restart the service.

I see where some additional notifys were added, but to be honest, a 10 second timeout seems too short anyway, and cutting it so close was a gamble.  So I raised the timeout, and while I did keep the notify in Files (although I did move it to what was really the heavy call), with the additional time, I don't really think it is nessesary.

This additional buffer, and adhering to the documented call cadence, should resolve the random reboots (I know 2.7.0 was supposed to resolve these, likely due to the addition of the additional notify calls, however I just had a development board doing nothing reset, so...).

Also, there was a Notify Ready in Files.  That is reserved for Service Init, so I pulled that out.