Fix the 'FatalErrorWindow' to... BE a Window...
The previous incantation of this process didn't really do anything of use.  It made a window, but, because it called 'LogFatalError', the program IMMEDIATLY quit, without displaying the error.

Beyond making it NOT display a window, if the error had something to do with configuration, this meant that the app was stuck in a restart loop that dumped stack traces to the log ever 2 seconds (unhelpful, and bad).

New behaviour actualy displays an error and leaves it on the screen.  The watchdog will timeout after however long, and the app will restart.  This makes more sense and makes the error display more useful.

TODO: This really should be located in a `defer` inside `main()`, which would allow it to handle PANIC.  Such an implimentation would kill the current `gtk` and start a whole new process, and would allow for 'throwing' errors anywhere in the application. This is outside the scope of the BF at this time, but MAY be added if this sits in PR for too long.